// src/features/employee-detail/domain/employee-detail.types.ts
export interface Employee {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  position: string;
  department: string;
  startDate: string;
  status: "active" | "inactive";
}

export interface Department {
  id: number;
  name: string;
}

export interface EmployeeFormData {
  firstName: string;
  lastName: string;
  email: string;
  position: string;
  department: string;
  startDate: string;
  status: "active" | "inactive";
}